+++ CHANGES.rst
@@ -1,5 +1,23 @@
+2.2.0
+-----
+The is a minor release:
+
+* Add support for Caja file manager (Linux).
+* Add option to select a specific file manager (Linux).
+* Remove unnecessary dependencies.
+* Add support for Dolphin file manager (Linux).
+
+2.1.0
+-----
+The is a minor release:
+
+* Add support for MacOSX.
+* Fix path to python for running scripts on gnome.
+
 2.0.0
 -----
+The is a major release changing the API:
+
 * Merge installation and removal command into one single commands
   ``start_jupyter_cm`` that takes a ``--remove`` argument.
 * Add test suite and continuous integration in travis and appveyor.